requirements.txt
openai-whisper==20240930
pyannote.audio==3.3.2
torch==2.2.0+cu121
pydub==0.25.1
scipy==1.15.3
tqdm==4.67.1
numpy==1.26.4
transformers==4.55.0
huggingface-hub==0.34.4
soundfile==0.13.1
einops==0.8.1
omegaconf==2.3.0
pytorch-lightning==2.5.2
lightning==2.5.2
tokenizers==0.21.4
tiktoken==0.11.0
pytest
scikit-learn==1.5.1